+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  * `Seed::add`
  * impl ops::{Add,Sub} for Coordinates
+ * `hexbot::WidthHeight` improved replacement for `WithCoordinates` and `CoordinatesLimit`
 
 ## [0.0.10] - 2019-09-28
 ### Changed